requirements.txt
# Modernized requirements for native Florence-2 and stable iopaint
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0
torchvision>=0.19.0

# transformers >= 4.45 includes native Florence-2 support
transformers>=4.45.0
# diffusers >= 0.30.0 avoids the hf_hub cached_download error
diffusers>=0.30.0
# huggingface-hub >= 0.23.0 is compatible with modern transformers/diffusers
huggingface-hub>=0.23.0,<1.0.0
peft>=0.12.0
numpy<2

# Image Processing
opencv-python-headless>=4.8.0,<4.12.0
Pillow>=10.0.0

# Utilities
loguru
click
tqdm
psutil
pyyaml

# iopaint helper dependencies (installed explicitly because iopaint is installed with --no-deps)
pydantic
typer
einops
omegaconf
easydict
yacs
timm
# used for Florence-2
einops
# bridge dependencies
filelock
fsspec
